+++ R/as_latex.R
@@ -6,9 +6,10 @@
 #' `data.frame`s) to latex code.
 #'
 #' `as_latex()` and co. are designed to produce nice looking output with a
-#' minimum of user input required. This is useful if you want a quick preview
-#' or printout of a table.  If you need customized Latex the output, you
-#' should take a look at the packages [kableExtra::kableExtra], **xtable**, or **huxtable**.
+#' minimum of user input required. This is useful if you want a quick preview or
+#' printout of a table.  If you need customized Latex the output, you should
+#' take a look at the packages [kableExtra::kableExtra], **xtable**, or
+#' **huxtable**.
 #'
 #' @section Latex Packages:
 #' `as_latex` requires that the following Latex packages are installed on your
@@ -377,7 +378,7 @@ multinames_to_colspans <- function(x){
 #' Default Kable options for as_latex and co
 #'
 #' `default_kable_options()` returns a list of the default options that are
-#' required for [`as_latex()`] to work correclty. Those defaults should not be
+#' required for [`as_latex()`] to work correctly. Those defaults should not be
 #' modified, but you can pass additional [knitr::kable()] options to
 #' `as_latex()` to modify the output a bit.
 #'