+++ R/tatoo-package.R
@@ -29,4 +29,11 @@
 
 
 
-if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
+if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", "sheet"))
+
+
+
+
+ignore_import_warnings <- function(...){
+  colt::clt_bg_accent("blubb")
+}